pyproject.toml
[tool.poetry]
name = "myfastapi"
version = "0.1.0"
description = ""
authors = ["thkim <thkim96@ssrinc.co.kr>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.5"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
py-ulid = "^1.0.3"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
sqlalchemy = "^2.0.37"
mysqlclient = "^2.2.7"
alembic = "^1.14.0"
dependency-injector = "^4.45.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

